+++ game/board.py
@@ -15,3 +15,9 @@ class Board:
 
     def get_piece(self,row,col):
         return self.__positions__[row][col]
+
+##hacer excepcion de lugar incorrecto
+    def move_piece(self, from_row, from_col, to_row, to_col):
+        piece = self.get_piece(from_row, from_col)
+        self.__positions__[from_row][from_col] = None
+        self.__positions__[to_row][to_col] = piece