+++ test/test_board.py
@@ -4,6 +4,10 @@ from game.rook import Rook
 
 
 class TestBoard(unittest.TestCase):
+
+    def setUp(self):
+        self.board = Board()
+
     def test_initial_positions_rook(self):
         board = Board()
         self.assertIsInstance(board.get_piece(0, 0), Rook)        
@@ -18,6 +22,14 @@ class TestBoard(unittest.TestCase):
                 if (row, col) not in [(0, 0), (0, 7), (7, 0), (7, 7)]:
                     self.assertIsNone(board.get_piece(row, col))
 
+    def test_move_piece(self):
+        self.board.move_piece(0, 0, 0, 1)
+        self.assertIsInstance(self.board.get_piece(0, 1), Rook)
+        self.assertIsNone(self.board.get_piece(0, 0))
+
+        self.board.move_piece(0, 1, 1, 1)
+        self.assertIsInstance(self.board.get_piece(1, 1), Rook)
+        self.assertIsNone(self.board.get_piece(0, 1))
 
 if __name__ == '__main__':
     unittest.main()